+++ pkg/controller/subnet.go
@@ -684,7 +684,7 @@ func (c *Controller) handleAddOrUpdateSubnet(key string) error {
 		return err
 	}
 
-	if value, ok := vpc.Annotations[util.VpcEnableLbAnnotation]; ok && value == util.VpcAnnotationEnableOn && c.config.EnableLb && subnet.Name != c.config.NodeSwitch {
+	if value, ok := vpc.Annotations[util.VpcEnableLbAnnotation]; ok && value == "true" && c.config.EnableLb && subnet.Name != c.config.NodeSwitch {
 		if err := c.ovnClient.AddLbToLogicalSwitch(vpc.Status.TcpLoadBalancer, vpc.Status.TcpSessionLoadBalancer, vpc.Status.UdpLoadBalancer, vpc.Status.UdpSessionLoadBalancer, subnet.Name); err != nil {
 			c.patchSubnetStatus(subnet, "AddLbToLogicalSwitchFailed", err.Error())
 			return err
