+++ pkg/util/const.go
@@ -161,7 +161,4 @@ const (
 	NetemQosLatencyAnnotationTemplate = "%s.kubernetes.io/latency"
 	NetemQosLimitAnnotationTemplate   = "%s.kubernetes.io/limit"
 	NetemQosLossAnnotationTemplate    = "%s.kubernetes.io/loss"
-
-	VpcAnnotationEnableOn  = "on"
-	VpcAnnotationEnableOff = "off"
 )
